refactor: extract community streak hook
